pyproject.toml
[tool.poetry]
name = "albus"
version = "0.1.0"
description = ""
authors = ["Marco Andronaco <andronacomarco@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pypandoc = "^1.13"
python-telegram-bot = "^21.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
